1.2
1.2

+++ src/util/uuid.h
@@ -22,7 +22,7 @@
 
 #undef uuid_t
 typedef struct {
-    unsigned long time_low;
+    unsigned int time_low;
     unsigned short time_mid;
     unsigned short time_hi_and_version;
     unsigned char clock_seq_hi_and_reserved;
